Use MkDocs for documentation generation
The documentation is in need of a redesign. Currently, the index is big,
badly placed, but worst of all: static.

MkDocs is a static site builder specificly designed for something like
the fikkie docs and it supports a clone of the readthedocs theme which
looks great! It looks way better this way and serving the docs locally
does not require anything other than Python-only packages. How nice.