Bump PyPi package to Beta status
Fikkie currently has every feature that was on the TODO list, so it's
time to bump the development status from Alpha to Beta.

The next milestone is v1.0.0. This will be reached when multiple users
have been using fikkie for a while, major bugs have been fixed (if
existing) and the API's have "matured". We don't want to reach v1.0.0
and discover we need a breaking change in the API! :)